Make point cloud directory optional in metadata script
Updated prepare_segmentation_metadata.py to allow the point cloud directory to be optional. If not provided or invalid, point cloud paths in the generated metadata will be set to null, and appropriate warnings or debug messages are logged. The CLI argument for the point cloud directory is now optional with a default of None.